[Update] Untested async image load/upload & few tweaks over general app architecture
Used asynctask for rendering the image URL from the background thread. Nothing has been tested yet and behaviour has only
been written to work from a theoretical standpoint. Some client and server changes have been made - like exposing the user id to the client and using it in the image upload to sync the image id and the user id (this will be changed because the image table doesn't exist anymore and the image props are handled in the user db); it is possible that no image props may be needed in the user table for server-side handling and that sync with image file names and user ids would be enough to render from a URL. Several architectural changes have
been scheduled to be done after the image upload (including DB design -
img path = path + user id -
client-side model property types, and omission of SharedPreferences as
user information storage in favour of an API call each login for the
latest info).